requirements.txt
numpy<2.0.0
pandas<2.0.0
scikit-surprise
matplotlib
seaborn
